data/cdc/requirements.txt
kafka-python==2.0.2
pandas==2.2.3
s3fs==2024.12.0
pyarrow==18.1.0